backend/requirements.txt
Flask==3.1.1
pymongo==4.13.1
python-dotenv==1.1.0
bcrypt==4.3.0
PyJWT==2.10.1
flask-cors==6.0.1